requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
ldap3>=2.9.1
mysql-connector-python>=8.0.0
